unit-tests-datasets/python/requirements.txt
pytest==7.1.2
GitPython==3.1.27